backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-core==4.8.1
sqlparse==0.5.3
urllib3==2.5.0
PyPDF2==3.0.1
google-generativeai==0.7.1
python-dotenv==1.0.1
